web_app/PostBee/requirements.txt
django
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pillow
six
django-rest-framework
django-rest-framework-simplejwt
django-filter
django-markdown